Extract _misc.py from _legacy.py
