+++ modules/helpers/__init__.py
@@ -31,4 +31,5 @@ from ._fonts import *  # noqa: F403
 from ._overlays import *  # noqa: F403
 from ._artifacts import *  # noqa: F403
 from ._plex_cache import *  # noqa: F403
+from ._misc import *  # noqa: F403
 from ._paths import *  # noqa: F403
